Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    double a,b;
+    float a,b;
     char n;
     scanf("%f %f %c",&a,&b,&n);
     switch(n){
